requirements.txt
cmsplugin-filer==0.10.1
Django==1.7.8
django-allauth==0.20.0
django-appconf==1.0.1
django-bootstrap3==6.1.0
django-classy-tags==0.5
django-cms==3.1.1
django-contact-form==1.1
django-durationfield==0.5.1
django-filer==0.9.11
django-mptt==0.6.1
django-polymorphic==0.7.1
django-reversion==1.8.0
django-sekizai==0.7
django-treebeard==3.0
djangocms-admin-style==0.2.5
djangocms-text-ckeditor==2.5.2
easy-thumbnails==2.2
html5lib==0.999
oauthlib==0.7.2
Pillow==2.8.2
psycopg2==2.6.1
python-openid==2.2.5
requests==2.7.0
requests-oauthlib==0.5.0
six==1.3.0
Unidecode==0.4.18
wheel==0.24.0
